feat: Update: 1.rename entity package
