+++ popcorn-scheduler/popcorn-scheduler-core/src/main/java/com/photowey/popcorn/scheduler/core/domain/entity/ScheduleApp.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.photowey.popcorn.scheduler.core.domain.enrity;
+package com.photowey.popcorn.scheduler.core.domain.entity;
 
 import java.io.Serializable;
